+++ Logos/src/utopia/logos/database/model/text/StatementLinkedModel.scala
@@ -7,7 +7,7 @@ import utopia.vault.model.immutable.Table
  * @author Mikko Hilpinen
  * @since 15.10.2023, Emissary Email Client v0.1, added to Logos v1.0 11.3.2024
  */
-trait StatementLinkModel
+trait StatementLinkedModel
 {
 	// ABSTRACT -----------------------